HBSD: Initial changes for HardenedBSD 12
We use the LiveCD builder for FuryBSD (FuryBSD is End-of-life) to build the desktop versions.

This is the initial version which is not perfect and still has bugs,
for example: the hardenedbsd kernel can't boot the ramdisk environment,
so you have to boot on the 2nd kernel (freebsd) to work around it temporarily.

Signed-off-by: Loic <loic.f@hardenedbsd.org>